introduce class PktLine for future Git protocol v2 support
Git protocol v2 adds a new pkt-line message separator which makes it
difficult to represent pkt-lines as either None or a bytestring.
The new PktLine class is able to represent pkt-lines for Git
protocol versions 0, 1, and 2.

The class supports a few convenience methods that allow existing code
to mostly work as-is. The bytestring is now stored in a 'data' member,
and PktLine provides some string-like methods to support existing code.
A PktLine can be cast to bytes(), which is useful when comparing the
pkt-line payload to a fixed sequence of bytes.

One major change is that a flush-pkt is no longer represented by the
None type. Instead, PktLine objects provide an is_flush_pkt() method.
Callers may compare the 'data' member to None to determine whether a
payload exists.